# Conflicts:
#	Controllers/ShiftsController.cs
#	Gride.sln
#	Migrations/20191114114702_initial.cs
#	Migrations/20191124191836_ShiftWeeklyBool.Designer.cs
#	Migrations/20191129103517_initial.Designer.cs
#	Migrations/20191129103517_initial.cs
#	Migrations/20191204143057_initial.Designer.cs
#	Migrations/20191204143057_initial.cs
#	Views/Shifts/Details.cshtml
#	Views/Shifts/Index.cshtml

+++ Controllers/WorksController.cs
@@ -7,51 +7,81 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Gride.Data;
 using Gride.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 
 namespace Gride.Controllers
 {
+    [Authorize]
     public class WorksController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly SignInManager<IdentityUser> signInManager;
+        private readonly UserManager<IdentityUser> userManager;
 
-        public WorksController(ApplicationDbContext context)
+        public WorksController(ApplicationDbContext context,
+                               SignInManager<IdentityUser> signInManager,
+                               UserManager<IdentityUser> userManager)
         {
             _context = context;
+            this.signInManager = signInManager;
+            this.userManager = userManager;
         }
 
         // GET: Works
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Works.Include(w => w.Employee).Include(w => w.Shift);
-            return View(await applicationDbContext.ToListAsync());
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
+            {
+                var applicationDbContext = _context.Works.Include(w => w.Employee).Include(w => w.Shift);
+                return View(await applicationDbContext.ToListAsync());
+            }
+            else
+            {
+                return Forbid();
+            }
         }
 
         // GET: Works/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var work = await _context.Works
-                .Include(w => w.Employee)
-                .Include(w => w.Shift)
-                .FirstOrDefaultAsync(m => m.WorkID == id);
-            if (work == null)
+                var work = await _context.Works
+                    .Include(w => w.Employee)
+                    .Include(w => w.Shift)
+                    .FirstOrDefaultAsync(m => m.WorkID == id);
+                if (work == null)
+                {
+                    return NotFound();
+                }
+
+                return View(work);
+            }
+            else
             {
-                return NotFound();
+                return Forbid();
             }
-
-            return View(work);
         }
 
         // GET: Works/Create
         public IActionResult Create(int ShiftId = 0)
         {
-            ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail");
-            ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", ShiftId); ;
-            return View();
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
+            {
+                ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail");
+                ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", ShiftId); ;
+                return View();
+            }
+            else
+            {
+                return Forbid();
+            }
         }
 
         // POST: Works/Create
@@ -61,33 +91,47 @@ namespace Gride.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("WorkID,EmployeeID,ShiftID,Overtime,Delay")] Work work)
         {
-            if (ModelState.IsValid)
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
             {
-                _context.Add(work);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(work);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail", work.EmployeeID);
+                ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", work.ShiftID);
+                return View(work);
             }
-            ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail", work.EmployeeID);
-            ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", work.ShiftID);
-            return View(work);
+            else
+            {
+                return Forbid();
+            }
         }
 
         // GET: Works/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var work = await _context.Works.FindAsync(id);
-            if (work == null)
-            {
-                return NotFound();
+                var work = await _context.Works.FindAsync(id);
+                if (work == null)
+                {
+                    return NotFound();
+                }
+                ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail", work.EmployeeID);
+                ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", work.ShiftID);
+                return View(work);
             }
-            ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail", work.EmployeeID);
-            ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", work.ShiftID);
-            return View(work);
+            else
+            {
+                return Forbid();
+            }
         }
 
         // POST: Works/Edit/5
@@ -97,54 +141,68 @@ namespace Gride.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("WorkID,EmployeeID,ShiftID,Overtime,Delay")] Work work)
         {
-            if (id != work.WorkID)
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
             {
-                return NotFound();
-            }
+                if (id != work.WorkID)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(work);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!WorkExists(work.WorkID))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(work);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!WorkExists(work.WorkID))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                return RedirectToAction(nameof(Index));
+                ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail", work.EmployeeID);
+                ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", work.ShiftID);
+                return View(work);
             }
-            ViewData["EmployeeID"] = new SelectList(_context.EmployeeModel, "ID", "EMail", work.EmployeeID);
-            ViewData["ShiftID"] = new SelectList(_context.Shift, "ShiftID", "ShiftID", work.ShiftID);
-            return View(work);
+            else
+            {
+                return Forbid();
+            }
         }
 
         // GET: Works/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
             {
-                return NotFound();
-            }
+                if (id == null)
+                {
+                    return NotFound();
+                }
 
-            var work = await _context.Works
-                .Include(w => w.Employee)
-                .Include(w => w.Shift)
-                .FirstOrDefaultAsync(m => m.WorkID == id);
-            if (work == null)
+                var work = await _context.Works
+                    .Include(w => w.Employee)
+                    .Include(w => w.Shift)
+                    .FirstOrDefaultAsync(m => m.WorkID == id);
+                if (work == null)
+                {
+                    return NotFound();
+                }
+
+                return View(work);
+            }
+            else
             {
-                return NotFound();
+                return Forbid();
             }
-
-            return View(work);
         }
 
         // POST: Works/Delete/5
@@ -152,10 +210,17 @@ namespace Gride.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var work = await _context.Works.FindAsync(id);
-            _context.Works.Remove(work);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            if (signInManager.IsSignedIn(User) && _context.EmployeeModel.Single(x => x.EMail == User.Identity.Name).Admin)
+            {
+                var work = await _context.Works.FindAsync(id);
+                _context.Works.Remove(work);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return Forbid();
+            }
         }
 
         private bool WorkExists(int id)
